+++ tests/test_parquet_export.py
@@ -1,3 +1,4 @@
+import os
 from io import BytesIO
 
 import pyarrow.parquet as pq
@@ -5,8 +6,9 @@ import pytest
 
 from udata_hydra.analysis.csv import (
     RESERVED_COLS,
+    csv_detective_routine,
+    csv_to_parquet,
     generate_records,
-    perform_csv_inspection,
 )
 from udata_hydra.utils.parquet import save_as_parquet
 
@@ -21,12 +23,12 @@ pytestmark = pytest.mark.asyncio
         ("catalog.xlsx", 2),
     ),
 )
-async def test_parquet_conversion(
-    setup_catalog, rmock, db, fake_check, produce_mock, file_and_count
-):
+async def test_save_as_parquet(file_and_count):
     filename, expected_count = file_and_count
     file_path = f"tests/data/{filename}"
-    inspection: dict | None = await perform_csv_inspection(file_path)
+    inspection: dict | None = csv_detective_routine(
+        csv_file_path=file_path, output_profile=True, num_rows=-1, save_results=False
+    )
     assert inspection
     columns = inspection["columns"]
     columns = {
@@ -41,3 +43,37 @@ async def test_parquet_conversion(
     assert len(table) == expected_count
     fake_file = BytesIO()
     pq.write_table(table, fake_file)
+
+
+@pytest.mark.parametrize(
+    "parquet_config",
+    (
+        (False, 1, False),  # CSV_TO_PARQUET = False, MIN_LINES_FOR_PARQUET = 1
+        (True, 1, True),  # CSV_TO_PARQUET = True, MIN_LINES_FOR_PARQUET = 1
+        (True, 3, False),  # CSV_TO_PARQUET = True, MIN_LINES_FOR_PARQUET = 3
+    ),
+)
+async def test_csv_to_parquet(mocker, parquet_config):
+    async def execute_csv_to_parquet() -> tuple[str, int] | None:
+        file_path = "tests/data/catalog.csv"
+        inspection: dict | None = csv_detective_routine(
+            csv_file_path=file_path, output_profile=True, num_rows=-1, save_results=False
+        )
+        assert inspection
+        return await csv_to_parquet(
+            file_path=file_path, inspection=inspection, table_name="test_table"
+        )
+
+    csv_to_parquet_config, min_lines_for_parquet_config, expected_conversion = parquet_config
+    mocker.patch("udata_hydra.config.CSV_TO_PARQUET", csv_to_parquet_config)
+    mocker.patch("udata_hydra.config.MIN_LINES_FOR_PARQUET", min_lines_for_parquet_config)
+
+    if not expected_conversion:
+        assert not await execute_csv_to_parquet()
+
+    else:
+        # TODO: don't use the exception as the assertion, better to mock the minio client sending the file
+        with pytest.raises(ValueError, match="invalid bucket name"):
+            await execute_csv_to_parquet()
+        # Clean the remaining parquet file
+        os.remove("test_table.parquet")